created charts for loss deployments

+++ rejects/models.py
@@ -6,9 +6,9 @@ class LossDeployment(models.Model):
     """Django data model LossDeployment"""
 
     shift = models.ForeignKey(ExtractLossData, on_delete=models.CASCADE, related_name='loss_deployment', verbose_name='Shift')
-    pkg_filler = models.IntegerField(blank=True, null=True, verbose_name='PKG Filler')
-    heuft_1_rejects = models.IntegerField(blank=True, null=True, verbose_name='Heuft 1 Rejects')
-    heuft_2_rejects = models.IntegerField(blank=True, null=True, verbose_name='Heuft 2 Rejects')
+    pkg_filler = models.DecimalField(max_digits=3, decimal_places=2, blank=True, null=True, verbose_name='PKG Filler')
+    heuft_1_rejects = models.DecimalField(max_digits=3, decimal_places=2,blank=True, null=True, verbose_name='Heuft 1 Rejects')
+    heuft_2_rejects = models.DecimalField(max_digits=3, decimal_places=2,blank=True, null=True, verbose_name='Heuft 2 Rejects')
 
     class Meta:
         verbose_name = 'LossDeployment'
